src/data_layer/connectors/bigquery.py
"""
BigQuery connector for production marketing data.

STUB: Implement this when you have access to GoFundMe's BigQuery.

The interface matches MockMarketingData exactly, so the rest of
your code works unchanged.
"""
from datetime import datetime
from typing import Any
import pandas as pd

from ..interfaces.base import BaseDataSource


class BigQueryMarketingData(BaseDataSource):
    """
    Production implementation using BigQuery.
    
    TODO: Fill in actual table names and queries when at GoFundMe.
    """
    
    def __init__(self):
        # Import here to avoid dependency issues in mock mode
        from google.cloud import bigquery
        from src.utils.config import settings
        
        self.client = bigquery.Client(project=settings.bq_project)
        self.dataset = settings.bq_dataset
        
        # TODO: Map channel names to actual BigQuery tables
        self.tables = {
            "google_search": f"{self.dataset}.google_search_performance",
            "google_pmax": f"{self.dataset}.google_pmax_performance",
            "meta_ads": f"{self.dataset}.meta_ads_performance",
            "tiktok_ads": f"{self.dataset}.tiktok_performance",
            # Add more mappings...
        }
    
    def get_metrics(
        self,
        channel: str,
        start_date: datetime,
        end_date: datetime,
        metrics: list[str] | None = None,
    ) -> pd.DataFrame:
        """Query BigQuery for channel metrics."""
        if channel not in self.tables:
            return pd.DataFrame()
        
        table = self.tables[channel]
        
        # Build column list
        if metrics:
            columns = ", ".join(["date"] + metrics)
        else:
            columns = "*"
        
        query = f"""
        SELECT {columns}
        FROM `{table}`
        WHERE date BETWEEN @start_date AND @end_date
        ORDER BY date
        """
        
        job_config = self.client.QueryJobConfig(
            query_parameters=[
                self.client.ScalarQueryParameter("start_date", "DATE", start_date.date()),
                self.client.ScalarQueryParameter("end_date", "DATE", end_date.date()),
            ]
        )
        
        return self.client.query(query, job_config=job_config).to_dataframe()
    
    def get_anomalies(
        self,
        channel: str | None = None,
        lookback_hours: int = 24,
        threshold_sigma: float = 2.0,
    ) -> list[dict[str, Any]]:
        """
        Query for anomalies.
        
        TODO: This could be a pre-computed table or a real-time query
        depending on GoFundMe's data architecture.
        """
        # Option 1: Query a pre-computed anomaly table
        # query = f"SELECT * FROM `{self.dataset}.anomalies_latest`"
        
        # Option 2: Compute on the fly (expensive but real-time)
        # For now, stub returns empty
        raise NotImplementedError("Implement based on GoFundMe's data architecture")
    
    def check_data_freshness(self) -> dict[str, datetime]:
        """Check when each table was last updated."""
        freshness = {}
        
        for channel, table in self.tables.items():
            query = f"""
            SELECT MAX(date) as last_date
            FROM `{table}`
            """
            result = self.client.query(query).to_dataframe()
            if not result.empty:
                freshness[channel] = result["last_date"].iloc[0]
        
        return freshness
    
    def is_healthy(self) -> bool:
        """Verify BigQuery connection."""
        try:
            self.client.query("SELECT 1").result()
            return True
        except Exception:
            return False
    
    def list_channels(self) -> list[str]:
        """List configured channels."""
        return list(self.tables.keys())
